Refactor day 2
Added files for each class/object that requires it.

Removed the level class, as it was never used and it was not complete

added lvl4 warnings and that warnings should be treated as errors and fixed all of those errors that occured

Fixed the insertnedhighscore function, where there was unnecesary work being done by the compiler

+++ Source/game.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "raylib.h"
 #include <vector>
-#include "Resources.h"
+#include "Resources.hpp"
 #include <string>
 
 //TODO: Remove redundant comments
@@ -76,7 +76,7 @@ public:
 	bool active; 
 	Color color; 
 	int health = 50;
-	int radius = 60;
+	float radius = 60;
 
 
 	void Render(Texture2D texture);
